move function_base out of ir/extension
