+++ paddle/cinn/ir/extension/CMakeLists.txt
@@ -1,2 +1,2 @@
 core_gather_headers()
-gather_srcs(cinnapi_src SRCS lowered_func.cc function_base.cc intrinsic_ops.cc)
+gather_srcs(cinnapi_src SRCS lowered_func.cc intrinsic_ops.cc)
